requirements.txt
Flask==3.1.0
mysql-connector-python==9.1.0
flask-wtf==1.2.2
flask-login==0.6.3
bcrypt==4.2.1
python-dotenv==1.0.1
flask-talisman==1.1.0
bleach==6.2.0
Flask-Limiter==3.8.0
